authorization via Gates and policies

+++ app/Http/Controllers/Api/EventController.php
@@ -15,15 +15,13 @@ class EventController extends Controller
 
     private array $relations = ['user', 'attendees', 'attendees.user'];
 
-    public function __construct()
-    {
-        Gate::authorizeResource(Event::class, 'event');
-    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        Gate::authorize('viewAny');
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -36,6 +34,7 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create',auth()->user());
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -54,6 +53,7 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
+        Gate::authorize('view', $event);
         return new EventResource($this->loadRelationships($event));
     }
 
@@ -62,7 +62,7 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
-//        Gate::authorize('update-event', $event);
+        Gate::authorize('update', $event);
         $event->update(
             $request->validate([
                 'name' => 'sometimes|string|max:255',
@@ -81,6 +81,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', $event);
         $event->delete();
         return response()->json(null, 204);
     }
